feat: finalização das implementações dos menus, melhoria no sistema de pastas do projeto
